another one

+++ Market Bot.py
@@ -19,14 +19,16 @@ async def check_shop_expire():
         print("checking for expired shops")
         remove_list = db.update_all_shop_durations()
         for shop in remove_list:
-            channel = get(bot.get_all_channels(), guild__name = 'Free Market', name = shop[0])
-            print("deleting " + channel.name)
-            await channel.delete()
-        announce_channel = get(bot.get_all_channels(), guild__name = 'Free Market', name = 'expired-shops-removal')
-        formatted_list = [f"{shop[0]:<30}{shop[1]:<30}" for shop in remove_list]
-        name, owner = "Name", "Owner"
-        msg = "The following shops are expired and have been removed\n"
-        await announce_channel.send(msg + '\n'.join([f"{name:<30}{owner:<30}"] + [""] + formatted_list))
+            for channel in bot.get_all_channels():
+                if channel.name == shop[0]:
+                    print("deleting " + channel.name)
+                    await channel.delete()
+        for channel in bot.get_all_channels():
+            if channel.name == 'expired-shops-removal':
+                formatted_list = [f"{shop[0]:<30}{shop[1]:<30}" for shop in remove_list]
+                name, owner = "Name", "Owner"
+                msg = "The following shops are expired and have been removed\n"
+                await channel.send(msg + '\n'.join([f"{name:<30}{owner:<30}"] + [""] + formatted_list))
         await asyncio.sleep(86400)
 
 @bot.event
